Add files via upload
This project on Olympic visualization using Tableau involves collecting and cleaning data related to the Olympic games, importing it into Tableau, and using it to create a comprehensive and interactive dashboard that provides an in-depth analysis of the Olympic games. This dashboard  includes visualizations such as medal counts, athlete analysis, event analysis, country analysis, and comparison analysis.